+++ src/main/java/com/github/sidhant92/boolparser/domain/arithmetic/ArithmeticNode.java
@@ -2,7 +2,7 @@ package com.github.sidhant92.boolparser.domain.arithmetic;
 
 import com.github.sidhant92.boolparser.constant.NodeType;
 import com.github.sidhant92.boolparser.constant.Operator;
-import com.github.sidhant92.boolparser.domain.Node;
+import com.github.sidhant92.boolparser.domain.logical.Node;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
