+++ src/main/java/com/github/sidhant92/boolparser/domain/arithmetic/UnaryNode.java
@@ -1,8 +1,7 @@
-package com.github.sidhant92.boolparser.domain;
+package com.github.sidhant92.boolparser.domain.arithmetic;
 
 import com.github.sidhant92.boolparser.constant.DataType;
 import com.github.sidhant92.boolparser.constant.NodeType;
-import com.github.sidhant92.boolparser.domain.arithmetic.ArithmeticBaseNode;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
