+++ src/main/java/com/github/sidhant92/boolparser/domain/logical/ArrayNode.java
@@ -1,6 +1,6 @@
-package com.github.sidhant92.boolparser.domain;
+package com.github.sidhant92.boolparser.domain.logical;
 
-import com.github.sidhant92.boolparser.constant.DataType;
+import java.util.List;
 import com.github.sidhant92.boolparser.constant.NodeType;
 import com.github.sidhant92.boolparser.constant.Operator;
 import lombok.AllArgsConstructor;
@@ -8,25 +8,18 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
-/**
- * @author sidhant.aggarwal
- * @since 05/03/2023
- */
 @AllArgsConstructor
 @Getter
 @Setter
 @Builder
-public class ComparisonNode extends Node {
+public class ArrayNode extends Node {
     private final String field;
-
-    private final Object value;
 
     private final Operator operator;
 
-    private final DataType dataType;
-
+    private final List<Node> items;
     @Override
     public NodeType getTokenType() {
-        return NodeType.COMPARISON;
+        return NodeType.ARRAY;
     }
 }
