+++ src/main/java/com/github/sidhant92/boolparser/domain/logical/InNode.java
@@ -1,10 +1,7 @@
-package com.github.sidhant92.boolparser.domain;
+package com.github.sidhant92.boolparser.domain.logical;
 
 import java.util.List;
-import org.apache.commons.lang3.tuple.Pair;
-import com.github.sidhant92.boolparser.constant.DataType;
 import com.github.sidhant92.boolparser.constant.NodeType;
-import com.github.sidhant92.boolparser.constant.Operator;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -14,14 +11,12 @@ import lombok.Setter;
 @Getter
 @Setter
 @Builder
-public class ArrayNode extends Node {
+public class InNode extends Node {
     private final String field;
-
-    private final Operator operator;
 
     private final List<Node> items;
     @Override
     public NodeType getTokenType() {
-        return NodeType.ARRAY;
+        return NodeType.IN;
     }
 }
