+++ src/main/java/com/github/sidhant92/boolparser/parser/BoolExpressionParser.java
@@ -1,6 +1,6 @@
 package com.github.sidhant92.boolparser.parser;
 
-import com.github.sidhant92.boolparser.domain.Node;
+import com.github.sidhant92.boolparser.domain.logical.Node;
 import io.vavr.control.Try;
 
 /**
